rag_eval_bdd/requirements.txt
pytest>=8.0.0
pytest-bdd>=7.0.0
allure-pytest>=2.13.5
requests>=2.31.0
pydantic>=2.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
deepeval>=1.5.0
rich>=13.7.1
matplotlib>=3.8.0
